+++ src/main/kotlin/dev/brunofelix/bookapi/repository/CustomerRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CustomerRepository : CrudRepository<Customer, Long> {
+interface CustomerRepository : CrudRepository<Customer, Int> {
     fun findByNameContaining(name: String): List<Customer>
 }
